# metal: Q3_K speedup