When running on a modern GPU using CUDA, `Q3_K` token generation (TG) is faster than `Q4_0` by a comfortable margin. On Metal, it is much slower than `Q4_0`. Assuming optimum implementation, TG is memory bound, so `Q3_K` should be always faster, given the smaller model size. I have tried various approaches, including the exact same matrix x vector multiplication kernel as on CUDA, to no avail.

This PR brings a ~15% speedup for `Q3_K` TG on Metal. It is still ~20% slower than `Q4_0`, but at least some progress.

Prompt processing is not affected by the PR, so table gives performance for TG only on a 30-core M2 Max:

| model                          | backend    | test       |    t/s (Master)  |     t/s (PR)  |  Speedup | 
| ------------------------------ | ---------- | ---------- | ---------------: | ------------: | -------: |
| LLaMA 7B mostly Q3_K - Small   | Metal      | tg 32      |     46.14 ± 0.01 |  52.77 ± 0.03 |  1.144   |
| LLaMA 7B mostly Q3_K - Small   | Metal      | tg 64      |     45.15 ± 0.55 |  52.60 ± 0.09 |  1.165   |
| LLaMA 7B mostly Q3_K - Small   | Metal      | tg 128     |     44.52 ± 0.05 |  52.26 ± 0.03 |  1.174   |   
| LLaMA 7B mostly Q3_K - Small   | Metal      | tg 256     |     44.19 ± 0.09 |  51.31 ± 0.24 |  1.161   |   
 
There is also a mini speedup for `Q5_K` (~2%).