@lshzh-ww 2023-09-04
Unfortunately, on the M1 Max, this kernel is slower than `master`. I suspect that this is still due to each thread directly reading from device memory. It is not a significant issue for the M2 series, but it affects M1 series performance because the M1 series has a weaker memory controller. Would you mind trying to first copy blocks in whole to threadgroup memory and then letting each thread read from there? I would be happy to test it for you.

| model                          |    backend    | ngl |    threads | test       |              t/s (Master) |          t/s (PR) | Speedup|
| ------------------------------ |---------- | --: | ---------: | ---------- | ---------------: |---------------: |---------: |
| llama-30b q3_K_S |  Metal      |   1 |          4 | tg 128     |  11.52 ± 0.01 |    10.49 ± 0.00 | 0.911 |

To address the issue of `Q3_K` being slower than `Q4_0`, I believe that `Q3_K` is actually compute-bound on M-series chips because they have lower FLOPS per memory bandwidth when compared to Nvidia GPUs.