@ggerganov 2023-09-04
- M2 Ultra

| model             |       size | test       |    master    t/s |   PR        t/s |   speedup   |
| ----------------- | ---------: | ---------- | ---------------: |---------------: |------------ |
| LLaMA 7B Q3_K_S   |   2.75 GiB | pp 32      |    568.81 ± 7.26 |   568.92 ± 6.53 |    1.000    |
| LLaMA 7B Q3_K_S   |   2.75 GiB | pp 64      |    763.14 ± 1.27 |   762.14 ± 1.39 |    0.999    |
| LLaMA 7B Q3_K_S   |   2.75 GiB | pp 128     |    888.56 ± 1.71 |   888.15 ± 1.96 |    1.000    |
| LLaMA 7B Q3_K_S   |   2.75 GiB | pp 256     |    907.66 ± 1.04 |   907.54 ± 0.72 |    1.000    |
| LLaMA 7B Q3_K_S   |   2.75 GiB | pp 512     |    876.53 ± 0.60 |   876.46 ± 0.17 |    1.000    |
| --- |
| LLaMA 7B Q3_K_S   |   2.75 GiB | tg 16      |     75.91 ± 0.08 |    83.43 ± 0.13 |    1.099    |
| LLaMA 7B Q3_K_S   |   2.75 GiB | tg 32      |     75.75 ± 0.03 |    83.31 ± 0.13 |    1.100    |
| LLaMA 7B Q3_K_S   |   2.75 GiB | tg 64      |     75.46 ± 0.04 |    83.11 ± 0.08 |    1.101    |
| LLaMA 7B Q3_K_S   |   2.75 GiB | tg 128     |     74.72 ± 0.07 |    82.15 ± 0.13 |    1.099    |
| LLaMA 7B Q3_K_S   |   2.75 GiB | tg 256     |     73.88 ± 0.06 |    81.20 ± 0.07 |    1.099    |


| model             |       size | test       |     master   t/s |     PR       t/s |   speedup   |
| ----------------- | ---------: | ---------- | ---------------: | ---------------: |------------ |
| LLaMA 30B Q3_K_S  |  13.10 GiB | tg 16      |   22.08 ± 0.01   |     24.62 ± 0.01 |  1.115      |
| LLaMA 30B Q3_K_S  |  13.10 GiB | tg 32      |   22.09 ± 0.01   |     24.62 ± 0.02 |  1.115      |
| LLaMA 30B Q3_K_S  |  13.10 GiB | tg 64      |   22.02 ± 0.02   |     24.57 ± 0.03 |  1.116      |
| LLaMA 30B Q3_K_S  |  13.10 GiB | tg 128     |   21.78 ± 0.02   |     24.38 ± 0.01 |  1.119      |


- M1 Pro

| model             |       size | test       |     master   t/s |     PR       t/s |   speedup   |
| ----------------- | ---------: | ---------- | ---------------: | ---------------: |------------ |
| llama2 7B Q3_K_S  |   2.75 GiB | tg 16      |     26.37 ± 0.04 |  28.00 ± 0.02    |   1.062     |
| llama2 7B Q3_K_S  |   2.75 GiB | tg 32      |     26.33 ± 0.01 |  27.96 ± 0.01    |   1.062     |
| llama2 7B Q3_K_S  |   2.75 GiB | tg 64      |     26.26 ± 0.01 |  27.87 ± 0.01    |   1.061     |
| llama2 7B Q3_K_S  |   2.75 GiB | tg 128     |     26.12 ± 0.01 |  27.74 ± 0.01    |   1.062     |
| llama2 7B Q3_K_S  |   2.75 GiB | tg 256     |     25.94 ± 0.01 |  27.54 ± 0.00    |   1.062     |


The M1 Pro also benefits from this change - at least for the small 7B model