@ikawrakow 2023-09-04
I don't see a big difference in performance improvement between 7B and 30B. The gain is ~12-14% vs ~14-17% for 7B:

| model                          | backend    | test       |     t/s (Master) |     t/s (PR)     |  Speedup |
| ------------------------------ | ---------- | ---------- | ---------------: | ---------------: | -------: |
| LLaMA 30B mostly Q3_K - Small  | Metal      | tg 32      |     11.38 ± 0.02 |     12.93 ± 0.03 | 1.136    |   
| LLaMA 30B mostly Q3_K - Small  | Metal      | tg 64      |     11.33 ± 0.00 |     12.85 ± 0.01 | 1.134    |   
| LLaMA 30B mostly Q3_K - Small  | Metal      | tg 128     |     11.20 ± 0.10 |     12.81 ± 0.04 | 1.144    |   
| LLaMA 30B mostly Q3_K - Small  | Metal      | tg 256     |     11.22 ± 0.01 |     12.57 ± 0.17 | 1.120    |