@ikawrakow 2023-09-04
@lshzh-ww 
> To address the issue of Q3_K being slower than Q4_0, I believe that Q3_K is actually compute-bound on M-series chips because they have lower FLOPS per memory bandwidth when compared to Nvidia GPUs.

If that was true, we wouldn't see the more than ~8X difference in t/s between TG and PP we are observing on the M series, which applies to all quantization types. On a modern GPU the performance difference is even more pronounced because of more compute being available (e.g., on my RTX-4080 TG-128 is ~130 t/s while PP-512 is ~3400 t/s, so 28X difference). On M2 `Q3_K` PP performance is ~85% of `Q4_0` PP performance, which is easily understandable considering how much more bit fiddling per quant is necessary for `Q3_K` compared to `Q4_0`. On the M series, if implemented optimally, TG ought to become memory bound, with perhaps some small influence of available compute performance.