@lshzh-ww 2023-09-04
> If that was true, we wouldn't see the more than ~8X difference in t/s between TG and PP we are observing on the M series, which applies to all quantization types. On a modern GPU the performance difference is even more pronounced because of more compute being available (e.g., on my RTX-4080 TG-128 is ~130 t/s while PP-512 is ~3400 t/s, so 28X difference).

I am sorry, but we shouldn't compare the performance between TG and PP to estimate the compute pressure for dequantization. In matrix-matrix multiplication kernels, we split the `dst` matrix into tiles of 64x32 dimensions. By doing this, we reduce the memory load pressure and computation needs for dequantization to 1/32. (i.e., each weight in `src0` is loaded once, dequantized once, and reused 32 times, compared to matrix-vector multiplication kernels where each weight in `src0` is loaded once, dequantized once, and used once.)

If this doesn't convince you, here are the profiling results for the `master` branch :).
<img width="300" alt="Screenshot 2023-09-04 at 11 32 52" src="https://github.com/ggerganov/llama.cpp/assets/61452103/5c635eea-5c4c-4cb3-b02a-1fb680237c3e">

@ggerganov 
M1 Max 32c GPU, 7B model:
| model                          |    backend    | ngl |    threads | test       |              t/s (Master) |          t/s (PR) | Speedup|
| ------------------------------ |---------- | --: | ---------: | ---------- | ---------------: |---------------: |---------: |
| codellama-7b q3_K_S |  Metal      |   1 |          4 | tg 128     |  45.95 ± 0.01 |    43.38 ± 0.00 | 0.944 |

By the way, are your results from the M1 Pro with 16 GPU cores or 14 GPU cores? Since `Q3_K` is more compute-bound, it may be worth collecting more benchmark results from various GPU core configurations.