@ggerganov 2023-09-04
@lshzh-ww 

It has 16 GPU cores. A 32GB  model, running macOS 13.4.1.

![image](https://github.com/ggerganov/llama.cpp/assets/1991296/6ec4e13f-5f9f-4664-8ded-cfa3a6932e58)