@ikawrakow 2023-09-07
@ggerganov What do we do with this? It is faster on M2 Max, M2 Ultra, M1 Pro, but somehow slower on M1 Max from 1 datapoint and no further feedback.